Refactor draggable_list.dart
- Refactor DraggableList to Category in the same file
- Refactor DraggableListItem to CategoryItem in a new file
- Refactor Categories to a new file
- Add comments to all the files
- Edit the files which use those datatypes to use the new names
