+++ lib/models/category.dart
@@ -1,28 +1,47 @@
-// This Model class serves to facilitate the interaction with records from the categories collection.
+import 'package:seg_coursework_app/models/category_item.dart';
 
-// It allows controlling and interacting with its fields through a special constructor that enables creating
-// a Collection object right from a database response. It also allows to easily export all manipulation to the
-// database by being easily convertible to a Map object that can then be passed on to Firebase.
+/// A datatype model which mimics a 'Category' entry in Firestore
 class Category {
+  final bool availability;
   final String title;
-  final String illustration;
-  final int page_rank;
+  final String imageUrl;
+  final String id;
+  final int rank;
+  final List<CategoryItem> items;
 
   Category(
       {required this.title,
-      required this.illustration,
-      required this.page_rank});
+      required this.rank,
+      required this.items,
+      required this.imageUrl,
+      required this.id,
+      required this.availability});
 
-  Map<String, dynamic> toMap() {
-    return {
-      'title': title,
-      'illustration': illustration,
-      'rank': page_rank,
-    };
+  /// Convert a Category back from json data
+  factory Category.fromJson(Map<String, dynamic> json) {
+    final items = (json['items'] as List<dynamic>)
+        .map((e) => CategoryItem.fromJson(e as Map<String, dynamic>))
+        .toList();
+    return Category(
+      title: json['title'] as String,
+      rank: json['rank'] as int,
+      items: items,
+      imageUrl: json['imageUrl'] as String,
+      id: json['id'] as String,
+      availability: json['availability'] as bool,
+    );
   }
 
-  Category.fromMap(Map<String, dynamic> addressMap)
-      : title = addressMap["title"],
-        illustration = addressMap["illustration"],
-        page_rank = addressMap["rank"];
+  /// Convert the current Category to json data
+  Map<String, dynamic> toJson() {
+    final items = this.items.map((e) => e.toJson()).toList();
+    return {
+      'title': title,
+      'rank': rank,
+      'items': items,
+      'imageUrl': imageUrl,
+      'id': id,
+      'availability': availability,
+    };
+  }
 }
